Revert compat function moves
